requirements.txt
numpy==1.20.2
matplotlib==3.4.1
tqdm==4.60.0
scipy==1.6.2
pandas==1.2.4
Pillow==8.2.0
